backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-generativeai
python-dotenv
pydantic
requests
python-multipart
Pillow
